+++ src/polygon.h
@@ -23,18 +23,18 @@ class Contour {
   typedef std::vector<Point_2>::const_iterator const_iterator;
 
   Contour();
-  Contour(const std::vector<Point_2>& points,
-          const std::vector<unsigned int>& holes, bool external);
+  Contour(const std::vector<Point_2>& points, const std::vector<size_t>& holes,
+          bool external);
 
   /** Get the p-th vertex of the external contour */
-  Point_2& vertex(unsigned int p) { return _points[p]; }
-  Point_2 vertex(unsigned int p) const { return _points[p]; }
-  Segment_2 segment(unsigned int p) const {
+  Point_2& vertex(size_t p) { return _points[p]; }
+  Point_2 vertex(size_t p) const { return _points[p]; }
+  Segment_2 segment(size_t p) const {
     return (p == nvertices() - 1) ? Segment_2(_points.back(), _points.front())
                                   : Segment_2(_points[p], _points[p + 1]);
   }
   /** Number of vertices and edges */
-  unsigned nvertices() const { return _points.size(); }
+  size_t nvertices() const { return _points.size(); }
   /** Get the bounding box */
   Bbox_2 bbox() const;
   /** Return if the contour is counterclockwise oriented */
@@ -62,9 +62,9 @@ class Contour {
   const_iterator end() const { return _points.end(); }
   Point_2& back() { return _points.back(); }
   const Point_2& back() const { return _points.back(); }
-  void addHole(unsigned int ind) { _holes.push_back(ind); }
-  unsigned int nholes() const { return _holes.size(); }
-  unsigned int hole(unsigned int p) const { return _holes[p]; }
+  void addHole(size_t ind) { _holes.push_back(ind); }
+  size_t nholes() const { return _holes.size(); }
+  size_t hole(size_t p) const { return _holes[p]; }
   bool external() const { return _external; }
   void setExternal(bool e) { _external = e; }
 
@@ -73,7 +73,7 @@ class Contour {
   std::vector<Point_2> _points;
   /** Holes of the contour. They are stored as the indexes of the holes in a
    * polygon class */
-  std::vector<unsigned int> _holes;
+  std::vector<size_t> _holes;
   bool _external;  // is the contour an external contour? (i.e., is it not a
                    // hole?)
   bool _CC;
@@ -93,14 +93,14 @@ class Polygon {
   bool open(const std::string& filename);
   void join(const Polygon& pol);
   /** Get the p-th contour */
-  Contour& contour(unsigned int p) { return _contours[p]; }
-  const Contour& contour(unsigned int p) const { return _contours[p]; }
-  Contour& operator[](unsigned int p) { return _contours[p]; }
-  const Contour& operator[](unsigned int p) const { return _contours[p]; }
+  Contour& contour(size_t p) { return _contours[p]; }
+  const Contour& contour(size_t p) const { return _contours[p]; }
+  Contour& operator[](size_t p) { return _contours[p]; }
+  const Contour& operator[](size_t p) const { return _contours[p]; }
   /** Number of contours */
-  unsigned int ncontours() const { return _contours.size(); }
+  size_t ncontours() const { return _contours.size(); }
   /** Number of vertices */
-  unsigned int nvertices() const;
+  size_t nvertices() const;
   /** Get the bounding box */
   Bbox_2 bbox() const;
   void push_back(const Contour& c) { _contours.push_back(c); }
